Add versioning and bump to 0.1.0 (#44)
This PR adds basic support for versioning, it let's a user check the version of the imported package with `examples_utils.__version__` and is set during installtion by pip in the  setup() for correct removal of stale python "eggs".


See the new output of a pip install call:

```
pip install examples-utils@...
...
Successfully installed examples-utils-0.1.0
```

The version is manually parsed from `__init__.py` because it cannot be imported without
making all the dependencies, install dependencies.

Tests were run and passed. Notebook tests check that the package can be installed: the new code is covered by the test suite.