Allow customisation of metrics and CSV reports from client code (#61)
This commit makes it possible to register parsing functions which will be
applied to the log output of a benchmark. This is implemented in the
custom_metrics.py which includes a usage example. The same usage
example is put to the test in test_custom_metrics.

These features are added to the standard benchmarks command under
the `--custom-metrics-files` argument.