kill process more robustly, use variant-specific timeout if specified (#122)
This PR makes 2 changes to the benchmarking script:
1. Allow timeouts at the benchmark-variant level. On Paperspace CI we
have seen the issue where some notebooks hang/get stuck, and we want to
set timeouts for these apps that are different from other apps.
Previously the benchmarking scripts had a global timeout that would
apply to all benchmark variants. I added logic to retrieve a
variant-specific timeout in `benchmark_dict` if it's present, and to
take the min of the variant-specific and global timeout.

2. Make the 'kill process' in the event of a timeout more robust. We've
seen cases where a particular notebook will hang during compilation (in
C++ somewhere), and the previous way of killing the process
`subprocess.Popen().kill` wasn't enough to kill the thread and so we
still kept looping forever. I made a change to first kill all child
processes of the parent process before killing the parent process, which
resolved this issue.

Testing:
Ran a test on Paperspace notebooks with a combination of no timeout,
benchmark variant timeout and global timeout, and ensured all were
encountered correctly.