add option to retry downloading failed files from s3 (#23)
We've seen in CI that we encounter some 'connection reset' errors when
downloading from s3. In those cases, it seems to affect a few files (<
5) at a time, but so far without affecting the result of the notebook
run. This may be due to overloading the endpoint, or DMZ network issues.
Stuart has made adjustments to the s3 proxy VM to try to mitigate. But
at the same time, this PR makes a change to the existing download script
to attempt a retry if there are failures.

Changes made
- There is a `max-retries` command line arg (default 1)
- `parallel_download_dataset_from_s3` accepts a `failed_files` arg which
is None by default. If not None, the function will only try to download
those files, rather than all the datasets in settings.yaml.
- Errors from each attempt are accumulated in a `all_attempts_errors`
which is what is raised if all attempts are unsuccessful.
- We only retry the download if there are failed files, not if there are
missing datasets.
- If there are failed files from an earlier attempt that are successful
in a later attempt, we don't raise the previous error. In these cases,
we'd still raise a RuntimeError only if a different **type** of error
occurred in the first attempt, like a missing dataset. If all attempts
fail then errors from all attempts are printed/raised.

Testing:
All the tests below are from inspecting the s3 download in
`prepare-datasets.sh`

- CI run (successful on first download attempt) -
https://github.com/graphcore/Gradient-Pytorch-private/actions/runs/6237782767
- CI run forcing retry download of 1 file -
https://github.com/graphcore/Gradient-Pytorch-private/actions/runs/6238473814/job/16934381394.
@payoto in the second CI run's logs I noticed the following lines, just
wanted to check if it was expected
```
NB 09a61: End point cannot be reached from current executor: http://10.194.16.183:8300/
NB 09a61: Validated endpoint: http://10.129.255.154:8000/
```

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: rahult-graphcore <rahult-graphcore@users.noreply.github.com>